+++ database/migrations/2016_03_06_162405_create_ventas_table.php
@@ -27,8 +27,8 @@ class CreateVentasTable extends Migration
             $table->increments('id');
             $table->integer('producto_id')->unsigned();
             $table->integer('cantidad');
-            $table->integer('precio_compra');
-            $table->integer('precio_venta')->nullable();
+            $table->double('precio_compra', 11, 3);
+            $table->double('precio_venta', 11, 3)->nullable();
             $table->integer('total')->nullable();
             $table->timestamps();
         });
